add javadocs, fix issues
