+++ src/test/resources/fabric.mod.json
@@ -8,6 +8,7 @@
   "initializers": [
     "net.fabricmc.fabric.block.FireMod",
     "net.fabricmc.fabric.colormapper.ColorProviderMod",
+    "net.fabricmc.fabric.config.ConfigLoadMod",
     "net.fabricmc.fabric.containers.ContainerMod",
     "net.fabricmc.fabric.containers.ContainerModClient",
     "net.fabricmc.fabric.events.ServerEventMod",
